(ТО-52) Добавлены методы для мониторинга новых регионов и снятие мониторинга с удаленных регионов
