+++ MapSmartObjects/Modules/PinList/PinListPresenter.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import CoreLocation
 
 protocol IPinListPresenter
 {
@@ -19,12 +20,21 @@ protocol IPinListPresenter
 final class PinListPresenter
 {
 	weak var pinListViewController: PinListViewController?
+	weak var mapViewController: MapViewController?
 	private let repository: IRepository
 	private let router: IPinListRouter
 
 	init(repository: IRepository, router: IPinListRouter) {
 		self.repository = repository
 		self.router = router
+	}
+	// метод для остановки мониторинга зоны когда пользователь его удаляет(надо добавить когда удаляется пин)
+	private func stopMonitoring(smartObject: SmartObject) {
+		guard let regions = mapViewController?.locationManeger.monitoredRegions else { return }
+		for region in regions {
+			guard let circusRegion = region as? CLCircularRegion, circusRegion.identifier == smartObject.name else { continue }
+			mapViewController?.locationManeger.stopMonitoring(for: circusRegion)
+		}
 	}
 }
 
@@ -43,6 +53,7 @@ extension PinListPresenter: IPinListPresenter
 	}
 
 	func removeSmartObject(at index: Int) {
+		stopMonitoring(smartObject: repository.getSmartObjects()[index])
 		repository.removeSmartObject(at: index)
 	}
 }
